BBRv3 bug fix: use ProbeBW_DOWN pacing_gain of 0.9 instead of 0.75
Rationale:
  https://datatracker.ietf.org/meeting/117/materials/slides-117-ccwg-bbrv3-algorithm-bug-fixes-and-public-internet-deployment-00#page=6

Note that the ProbeBW_DOWN pacing_gain is already 0.9 in the doc,
because this particular BBRv3 bug fix had already been incorporated in
the previous draft revision,
draft-cardwell-iccrg-bbr-congestion-control-02. And the rationale was
already described, as well: "The pacing_gain value of 0.9 is derived
based on the ProbeBW_UP pacing gain of 1.25, as the minimum
pacing_gain value that allows bandwidth-based convergence to
approximate fairness."

But we touch these lines in the file to make it easier for folks to
find BBRv3 changes in the git commit history.