BBRv3 tuning: BBRStartupFullLossCnt=6
Rationale:
  https://datatracker.ietf.org/meeting/117/materials/slides-117-ccwg-bbrv3-algorithm-bug-fixes-and-public-internet-deployment-00#page=8

To trigger exit of Startup based on packet loss, BBRv3 requires
BBRStartupFullLossCnt loss events in a single round trip. For
loss-based exit of Startup, BBRv3 currently tunes the
BBRStartupFullLossCnt so that at least BBRStartupFullLossCnt=6
discontiguous sequence ranges that must be lost in that round trip to
exit Startup. Probably this will need further discussion, but I want
this to match the Linux TCP BBRv3 implementation for now.