Replace C.pipe with synonym C.inflight and clarify C.inflight
Replace occurrences of C.pipe with the synonym C.inflight, and clarify the definition of C.inflight.

Note that "pipe" is from RFC6675. During the recent PRRbis
review process we received the reasonable comment
that using  "pipe" is incorrect (or at least 
confusing/misleading) if the implementation
is not using RFC6675 for loss detection.

So let's standardize on "C.inflight", which is consistent with
PRRbis.