Replace C.pipe with synonym C.inflight and clarify C.inflight (#53)
Replace occurrences of C.pipe with the synonym C.inflight, and clarify the definition of C.inflight.

Note that "pipe" is from RFC6675. During the recent PRRbis
review process we received the reasonable comment
that using  "pipe" is incorrect (or at least 
confusing/misleading) if the implementation
is not using RFC6675 for loss detection.

So let's standardize on "C.inflight", which is consistent with
PRRbis.

* accept Ian's reword: "and have not been acknowledged or marked as lost since their last transmission"

Co-authored-by: ianswett <ianswett@users.noreply.github.com>

* accept Ian's edit: "C.cwnd = C.inflight + 1"

Co-authored-by: ianswett <ianswett@users.noreply.github.com>

* accept Ian's "reduce C.inflight and drain"

Co-authored-by: ianswett <ianswett@users.noreply.github.com>