clarify that BBR doc uses units of bytes, sec, and bytes/sec (#85)
* clarify that BBR doc uses units of bytes, sec, and bytes/sec

Clarify near the start of the "Terminology" section that
this document uses units of bytes, sec, and bytes/sec.

Clarify that implementations MAY use other units, with
appropriate conversions (since Linux TCP uses packets).

But note that it is RECOMMENDED to use bytes rather than
packets, since packet sizes can vary.

Try to remove mentions of units in the main sections of the
document. Where it's more clear to use units, make sure
the text is consistent with the bytes, sec, and bytes/sec units.

To keep the doc more succinct and readable, and avoid churn,
and avoid cognitive dissonance with respect to
implementations using other units (like Linux TCP), we still
avoid putting units in the names of variables in this doc.

Fixes #27

* refine "the sender knows the exact bytes that were delivered"

Given TCP retransmission ambiguity, it's not clear if at TCP
sender really knows "the exact bytes" that were delivered, in the sense of the identity of the bytes or packets.

And given that the L4 sender may not know how many bytes of
headers or encapsulation are added, it's not strictly true that
the sender always knows the "exact bytes" in the sense of
the exact number of bytes delivered at L1.

* from @ianswett: remove "(in bytes)"

Co-authored-by: ianswett <ianswett@users.noreply.github.com>

* from jbeshay: fix "bits persecond"

Co-authored-by: Joseph Beshay <jbeshay@meta.com>

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>
Co-authored-by: Joseph Beshay <jbeshay@meta.com>